ZEN changes for ZOPNET
Use the RZA path for ZEN sends. For ACKs/NACKs, message IDs are
not required, similar to a RESP ZOP from the RZA. For MZOPs/HZOPs
to the RZA, the ZEN internally manages a message id on both sends
and RESP receives.

Message sends from the ZAP use the existing message ID workflows like
other ZOPs. On receive in the ZEN, enqueue a message in the ZEN queue
by triggering the handler.