Testing in progress...
Added the ring to the forza_zqm_mbox_setup test (use the new onezap-onezone python); initializes, but doesn't seem to be computing the next address properly when trying to send the initial message from zap0 - is returning a nextAddress of 0 (which I shouldn't since the zap address is 0).

Once that's solved, we should be able to see the message start to go around the ring.  I hope.